requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.7  # For PostgreSQL on Render; optional for local SQLite
Flask-WTF>=1.0
python-dotenv>=0.21